# Add deferred decoding support for dictionary values