Closes #844
Closes #924

## Description

$subject.

### Performance
Performance for a dictionary with `String` keys and `String` values (`{String:String}`) of size 100:
```
name                                             old time/op    new time/op    delta
DictionaryDeferredDecoding/Simply_decode-12        57.4µs ±19%     3.2µs ± 1%  -94.48%  (p=0.016 n=5+4)
DictionaryDeferredDecoding/Get_value-12            50.2µs ± 9%    56.1µs ±29%     ~     (p=0.222 n=5+5)
DictionaryDeferredDecoding/Re-encode_decoded-12    14.0µs ±10%     1.8µs ± 7%  -87.20%  (p=0.008 n=5+5)

name                                             old alloc/op   new alloc/op   delta
DictionaryDeferredDecoding/Simply_decode-12        27.9kB ± 0%     0.5kB ± 0%  -98.19%  (p=0.008 n=5+5)
DictionaryDeferredDecoding/Get_value-12            27.9kB ± 0%    28.1kB ± 0%   +0.75%  (p=0.008 n=5+5)
DictionaryDeferredDecoding/Re-encode_decoded-12    6.47kB ± 0%    3.46kB ± 0%  -46.54%  (p=0.008 n=5+5)

name                                             old allocs/op  new allocs/op  delta
DictionaryDeferredDecoding/Simply_decode-12           622 ± 0%         6 ± 0%  -99.04%  (p=0.008 n=5+5)
DictionaryDeferredDecoding/Get_value-12               623 ± 0%       626 ± 0%   +0.48%  (p=0.008 n=5+5)
DictionaryDeferredDecoding/Re-encode_decoded-12      17.0 ± 0%      10.0 ± 0%  -41.18%  (p=0.008 n=5+5)
```
______

<!-- Complete: -->

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels